기존 Ladder 관련 class 새로운 class로 교체 완료
