--- src/main/java/ladder/domain/ladder/Cross.java
@@ -1,10 +1,0 @@
-package ladder.domain.ladder;
-
-public enum Cross {
-    CROSS,
-    NOT_CROSS;
-
-    boolean isCross() {
-        return (CROSS == this);
-    }
-}
