--- src/main/java/ladder/domain/ladder/generator/PointGenerator.java
@@ -1,9 +1,0 @@
-package ladder.domain.ladder.generator;
-
-import ladder.domain.ladder.Point;
-
-public interface PointGenerator {
-    Point generate(Point previous);
-    Point generateLeftmost();
-    Point generateRightmost(Point previous);
-}
