+++ src/test/java/ladder/domain/LadderGameTest.java
@@ -12,7 +12,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Arrays;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
@@ -33,8 +32,8 @@ public class LadderGameTest {
 
         LadderGameInfo ladderGameInfo = new LadderGameInfo(memberGroup, rewards);
 
-        Line2 line = new Line2(Arrays.asList(newPoint(0, false, false), newPoint(1, false, true), newPoint(2, true, false)));
-        Ladder2 ladder = new Ladder2(new Lines2(Arrays.asList(line)));
+        Line line = new Line(Arrays.asList(newPoint(0, false, false), newPoint(1, false, true), newPoint(2, true, false)));
+        Ladder ladder = new Ladder(new Lines(Arrays.asList(line)));
 
         // when
         // then
@@ -56,8 +55,8 @@ public class LadderGameTest {
 
         LadderGameInfo ladderGameInfo = new LadderGameInfo(memberGroup, rewards);
 
-        Line2 line = new Line2(Arrays.asList(newPoint(0, false, true), newPoint(1, true, false)));
-        Ladder2 ladder = new Ladder2(new Lines2(Arrays.asList(line)));
+        Line line = new Line(Arrays.asList(newPoint(0, false, true), newPoint(1, true, false)));
+        Ladder ladder = new Ladder(new Lines(Arrays.asList(line)));
 
         LadderGame ladderGame = new LadderGame(ladderGameInfo, ladder);
 
@@ -85,8 +84,8 @@ public class LadderGameTest {
 
         LadderGameInfo ladderGameInfo = new LadderGameInfo(memberGroup, rewards);
 
-        Line2 line = new Line2(Arrays.asList(newPoint(0, false, true), newPoint(1, true, false)));
-        Ladder2 ladder = new Ladder2(new Lines2(Arrays.asList(line)));
+        Line line = new Line(Arrays.asList(newPoint(0, false, true), newPoint(1, true, false)));
+        Ladder ladder = new Ladder(new Lines(Arrays.asList(line)));
 
         LadderGame ladderGame = new LadderGame(ladderGameInfo, ladder);
 
@@ -116,8 +115,8 @@ public class LadderGameTest {
 
         LadderGameInfo ladderGameInfo = new LadderGameInfo(memberGroup, rewards);
 
-        Line2 line = new Line2(Arrays.asList(newPoint(0, false, true), newPoint(1, true, false)));
-        Ladder2 ladder = new Ladder2(new Lines2(Arrays.asList(line)));
+        Line line = new Line(Arrays.asList(newPoint(0, false, true), newPoint(1, true, false)));
+        Ladder ladder = new Ladder(new Lines(Arrays.asList(line)));
 
         LadderGame ladderGame = new LadderGame(ladderGameInfo, ladder);
 
@@ -137,8 +136,8 @@ public class LadderGameTest {
         return new Coordinate(new CoordinateValue(x), new CoordinateValue(y));
     }
 
-    private Point2 newPoint(int index, boolean left, boolean right) {
+    private Point newPoint(int index, boolean left, boolean right) {
         Direction direction = new Direction(left, right);
-        return new Point2(index, direction);
+        return new Point(index, direction);
     }
 }
