+++ src/test/java/ladder/domain/ladder/LineTest.java
@@ -1,97 +1,102 @@
 package ladder.domain.ladder;
 
-import ladder.vo.coordinate.Coordinate;
-import ladder.vo.coordinate.CoordinateValue;
+import ladder.domain.ladder.generator.DirectionGenerator;
+import ladder.domain.ladder.generator.StayDirectionGenerator;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Arrays;
-import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
+import static org.assertj.core.api.Assertions.*;
 
 public class LineTest {
     private static final Logger log = LoggerFactory.getLogger(LineTest.class);
 
     @Test
-    public void Line의_Point가_하나일_경우_IllegalArgumentException() {
-        // (참가자가 한 명일 경우)
-        // given
-        Point point = Point.LEFT_CROSS_POINT;
-
+    public void 생성_시_사이즈가_2_미만일_경우_IllegalArgumentException() {
         // when
         // then
-        assertThatIllegalArgumentException().isThrownBy(() -> new Line(Arrays.asList(point)));
+        DirectionGenerator directionGenerator = new StayDirectionGenerator();
+        assertThatIllegalArgumentException().isThrownBy(() -> Line.init(1, directionGenerator));
     }
 
     @Test
-    public void 가장_왼쪽의_Point가_왼쪽으로_건너갈_수_있는_경우_IllegalArgumentException() {
+    public void 첫번째_Point의_left가_true면_IllegalArgumentException() {
         // given
-        Point point = Point.LEFT_CROSS_POINT;
+        Point previous = new Point(0, new Direction(true, false));
+        Point next = new Point(1, new Direction(false, false));
 
         // when
         // then
-        assertThatIllegalArgumentException().isThrownBy(() -> new Line(Arrays.asList(point, point)));
+        assertThatIllegalArgumentException().isThrownBy(() -> new Line(Arrays.asList(previous, next)));
     }
 
     @Test
-    public void 가장_오른쪽의_Point가_오른쪽으로_건너갈_수_있는_경우_IllegalArgumentException() {
+    public void 마지막_Point의_right가_true면_IllegalArgumentException() {
         // given
-        Point point = Point.RIGHT_CROSS_POINT;
+        Point previous = new Point(0, new Direction(false, false));
+        Point next = new Point(1, new Direction(false, true));
 
         // when
         // then
-        assertThatIllegalArgumentException().isThrownBy(() -> new Line(Arrays.asList(point, point)));
+        assertThatIllegalArgumentException().isThrownBy(() -> new Line(Arrays.asList(previous, next)));
     }
 
     @Test
-    public void 인접한_Point_끼리_Cross가_일치하지_않는_경우_IllegalArgumentException() {
+    public void 생성_시_이전_Point와_다음_Point가_맞물리지_않으면_IllegalStateException() {
+        // given
+        Point previous = new Point(0, new Direction(false, false));
+        Point next = new Point(1, new Direction(true, false));
+
         // when
         // then
-        assertThatIllegalArgumentException().isThrownBy(() ->
-                new Line(Arrays.asList(Point.RIGHT_CROSS_POINT, Point.LEFT_CROSS_POINT, Point.LEFT_CROSS_POINT)));
+        assertThatIllegalStateException().isThrownBy(() -> new Line(Arrays.asList(previous, next)));
     }
 
     @Test
-    public void Line에서_좌표_좌우로_넘어가기() {
+    public void 정상_생성() {
         // given
-        List<Point> points = Arrays.asList(Point.RIGHT_CROSS_POINT, Point.LEFT_CROSS_POINT, Point.CANNOT_CROSS_POINT);
-        Line line = new Line(points);
-
-        Coordinate firstCoordinate = getCoordinate(0, 0);
-        Coordinate secondCoordinate = getCoordinate(1, 0);
-        Coordinate thirdCoordinate = getCoordinate(2, 0);
+        int initialSize = 2;
+        DirectionGenerator directionGenerator = new StayDirectionGenerator();
 
         // when
-        Coordinate shouldBeSecond = line.cross(firstCoordinate);
-        Coordinate shouldBeFirst = line.cross(secondCoordinate);
-        Coordinate shouldBeThird = line.cross(thirdCoordinate);
+        Line line = Line.init(initialSize, directionGenerator);
 
         // then
-        assertThat(shouldBeFirst).isEqualTo(getCoordinate(0, 0));
-        assertThat(shouldBeSecond).isEqualTo(getCoordinate(1, 0));
-        assertThat(shouldBeThird).isEqualTo(getCoordinate(2, 0));
+        assertThat(line.getSize()).isEqualTo(initialSize);
 
         log.debug("line\n{}", line);
+    }
+
+    @Test
+    public void Line_내에서_이동() {
+        // given
+        Point previous = new Point(0, new Direction(false, true));
+        Point next = new Point(1, new Direction(true, false));
+
+        Line line = new Line(Arrays.asList(previous, next));
+
+        // when
+        int shouldBeOne = line.move(0);
+
+        // then
+        assertThat(shouldBeOne).isEqualTo(1);
     }
 
     @Test
     public void view로_보이는_String_테스트() {
         // given
-        List<Point> points = Arrays.asList(Point.RIGHT_CROSS_POINT, Point.LEFT_CROSS_POINT, Point.CANNOT_CROSS_POINT);
+        Point first = new Point(0, new Direction(false, false));
+        Point middle = new Point(1, new Direction(false, true));
+        Point last = new Point(2, new Direction(true, false));
 
         // when
-        Line line = new Line(points);
+        Line line = new Line(Arrays.asList(first, middle, last));
 
         // then
-        assertThat(line.toString()).contains("|-----|     |");
+        assertThat(line.toString()).contains("|     |-----|");
 
         log.debug("line\n{}", line);
-    }
-
-    private Coordinate getCoordinate(int x, int y) {
-        return new Coordinate(new CoordinateValue(x), new CoordinateValue(y));
     }
 }
